FuentesIliGutData

Fuentes S, et al. (2021) data in which the authors investigated associations between gut microbiota in respiratory infections and local and systemic immune markers in older adults during an influenza season.

data(FuentesIliGutData)

Format

Phyloseq object

Value

Loads the data set in R.

Details

FuentesIliGutData consists of 905 taxa (ASV-level) in 589 samples from 397 participants. The samples include follow-up (recovery) samples that were collected 7–9 weeks after the ILI event. Data on co-morbidities, medications, symptoms, pathogens are made available.

References

Fuentes S, et al. (2021) Associations of faecal microbiota with influenza-like illness in participants aged 60 years or older: an observational study. The Lancet Healthy Longevity, 2(1):e13-23. https://doi.org/10.1016/S2666-7568(20)30034-9

Ferreira JA, Fuentes S. (2020). Some comments on certain statistical aspects of the study of the microbiome. Briefings in bioinformatics 21(4), pp.1487-1494. https://doi.org/10.1093/bib/bbz077

Author

Sudarshan Shetty sudarshanshetty9@gmail.com