Neighborhood Classification Dysbiosis Scores

Neighborhood Classification Dysbiosis Scores

dysbiosisScore(x, reference_samples = NULL, method = c("median-BC", "BC-ED"))

Arguments

x

A phyloseq object

reference_samples

Vector of samples to use as reference.

method

Either 'median-BC' or 'BC-ED'

Value

A tibble data frame with dysbiosis score and sample information

Details

Calculates neighborhood classification based microbiome dysbiosis scores.

Dysbiosis score or index are often used to identify the extent to which clinically classified diseased samples are different from non diseased reference samples. There are several approaches large-scale bacterial marker profiling, relevant taxon-based methods, neighborhood classification, random forest prediction, and combined alpha-beta diversity reviewed by Wei 2021. Here we provide two scores based on neighborhood classification.

'median-BC' The median Bray-Curtis distance between the test sample and the reference samples (Lloyd-Price J, Arze C, Ananthakrishnan AN et al. 2021).

'BC-ED' The Euclidian distance between test sample and the centroid (median) of PCoA Axis 1 and 2 (based on Bray-Curtis dissimilarity) of the reference samples. (Mahapatra S et al 2021)

References

Lloyd-Price J, Arze C, Ananthakrishnan AN et al. (2019). Multi-omics of the gut microbial ecosystem in inflammatory bowel diseases. Nature, 569(7758), pp.655-662.

Mahapatra S et al. (2021). Nanopore 16S rRNA sequencing reveals alterations in nasopharyngeal microbiome and enrichment of Mycobacterium and Mycoplasma in patients with COVID 19. medRxiv preprint, 2021.11.10.21266147.

Author

Sudarshan A. Shetty

Examples

library(biomeUtils)
data("FuentesIliGutData")
# Keep only two groups. controls and L1.
ps <- filterSampleData(FuentesIliGutData, ILI != "L2")
ps <- getProportions(ps)
# Define controls as reference samples
ref_samples <- rownames(meta(subset_samples(ps, ILI == "C")))

dys_bc <- dysbiosisScore(ps,
                         reference_samples = ref_samples,
                         method="median-BC")
dys_bc_ed <- dysbiosisScore(ps,
                            reference_samples = ref_samples,
                            method="BC-ED")