Microbiota Uniqueness

Microbiota Uniqueness

uniqueness(x, dist_mat, reference_samples = NULL, ...)

Arguments

x

A phyloseq object

dist_mat

An object of class dist

reference_samples

Vector of samples to use as reference.

...

Option to pass biomeUtils::getSampleTibble like column name and columns to be included in the output.

Value

A data frame with uniqueness and sample information

Details

Extracts the minimum value from a dissimilarity matrix for each individual. This is the dissimilarity of an individual from their nearest neighbor. Here, the option of using a one or more reference samples is provided.

References

Wilmanski T et al. (2021). Gut microbiome pattern reflects healthy ageing and predicts survival in humans. Nature metabolism, 3(2), pp.274-286.

Author

Sudarshan A. Shetty

Examples

library(biomeUtils)
data("FuentesIliGutData")
# Keep only two groups. controls and L1.
ps <- filterSampleData(FuentesIliGutData, ILI != "L2")
ps <- getProportions(ps)
dist.mat <- phyloseq::distance(ps, "bray")
# Define controls as reference samples
ref_samples <- rownames(meta(subset_samples(ps, ILI == "C")))
muniq <- uniqueness(ps,
                    dist_mat=dist.mat,
                    reference_samples = ref_samples)
head(muniq)
#> # A tibble: 6 × 63
#>   S2     uniqueness Duplo..1.2. QPCR..fg.ul. Run.nr Reads.after.pip… sex     age
#>   <chr>       <dbl>       <int>        <dbl> <chr>             <dbl> <fct> <dbl>
#> 1 sampl…      0.428           1     45700000 272               14826 M        72
#> 2 sampl…      0.450           1     75100000 272               13754 F        68
#> 3 sampl…      0.502           1     27800000 273               47863 M        70
#> 4 sampl…      0.474           1      4300000 273               38995 M        78
#> 5 sampl…      0.597           1      3690000 273               22720 M        72
#> 6 sampl…      0.627           1     22200000 273               25059 F        61
#> # … with 55 more variables: ILI <fct>, BMI <dbl>, flu.vaccine <fct>,
#> #   respiratory.disease <fct>, cardiovascular.disease <fct>, diabetes <fct>,
#> #   renal.disease <fct>, transplant <fct>, autoimmune.disease <fct>, hiv <fct>,
#> #   asplenia <fct>, leukemia <fct>, lymphatic.cancer <fct>,
#> #   other.malignancies <fct>, smoking <fct>, passive.smoking <fct>,
#> #   statins <fct>, ACE.inhibitor <fct>, PPI <fct>, alphabeta.blocker <fct>,
#> #   antimicrobial <fct>, ARB <fct>, insuline <fct>, corticosteroids <fct>, …