Cluster ASVs into Clusters

Cluster ASVs into Clusters

Arguments

x
A phyloseq object

cluster_method
See methods in IdClusters in DECIPHER Default is "inexact"

cluster_cutoff
Cut-off for clustering ASVs. Default= 0.01 i.e. 99 percent identity

threads
No. of processors to use

Value

A new phyloseq object with cluster ASVs

Details

clusterASVs clusters ASV sequences obtained after dada2 in phyloseq object into clusters based on user provided sequence identity. If using method="inexact" this is similar to OTU approach for example clustering of ASVs that are 97 for ASVs or OTUs is highly debated (see reference section), both have their pros and cons.The user is advised to make appropriate choice based on their research questions and properties of data.

References

Clustering ASVs after DADA2? generating new OTU table. GitHub issue #947: https://github.com/benjjneb/dada2/issues/947

Michael McLaren. mikemc/speedyseq: speedyseq v0.2.0. Zenodo, 2020, June 30, (Version v0.2.0). http://doi.org/10.5281/zenodo.3923184

PD Schloss. Pre-print Review of 97 Blogpost: The Academic Hermit, 2017, Oct 11. http://www.academichermit.com/2017/10/11/Review-of-Updating-the-97-identity-threshold.html

Angela Oliverio and Noah Fierer. Intragenomic heterogeneity and its implications for ESVs. Blogpost: Fierer Lab, 2017, Oct 9. http://fiererlab.org/2017/10/09/intragenomic-heterogeneity-and-its-implications-for-esvs/

Noah Fierer, Tess Brewer, and Mallory Choudoir. Lumping versus splitting – is it time for microbial ecologists to abandon OTUs?. Blogpost: Fierer Lab, 2017, May 2. http://fiererlab.org/2017/05/02/lumping-versus-splitting-is-it-time-for-microbial-ecologists-to-abandon-otus/

Author

Sudarshan A. Shetty

Examples

library(biomeUtils)
library(dplyr)
library(data.table)

#> 
#> Attaching package: 'data.table'

#> The following objects are masked from 'package:dplyr':
#> 
#>     between, first, last

library(tibble)

testSprockett <- core(subset_samples(SprockettTHData, Community == "River_4"), 5, 20 / 100)
testSprockett <- clusterASVs(testSprockett,
                             cluster_method = "NJ",
                             cluster_cutoff = 0.01,
                             threads = 2)

#> Determining distance matrix based on shared 8-mers:
#> ================================================================================
#> 
#> Time difference of 0.03 secs
#> 
#> Clustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.02 secs
#> 
#> Aligning Sequences:
#> ================================================================================
#> 
#> Time difference of 1.75 secs
#> 
#> Iteration 1 of 2:
#> 
#> Determining distance matrix based on alignment:
#> ================================================================================
#> 
#> Time difference of 0.01 secs
#> 
#> Reclustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.02 secs
#> 
#> Realigning Sequences:
#> ================================================================================
#> 
#> Time difference of 1.11 secs
#> 
#> Iteration 2 of 2:
#> 
#> Determining distance matrix based on alignment:
#> ================================================================================
#> 
#> Time difference of 0.01 secs
#> 
#> Reclustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.02 secs
#> 
#> Realigning Sequences:
#> ================================================================================
#> 
#> Time difference of 0.28 secs
#> 
#> ================================================================================
#> 
#> Time difference of 0.01 secs
#> 
#> ================================================================================
#> 
#> Time difference of 0.01 secs
#>